requirements.txt
openai>=1.25.0
python-dotenv>=1.0.1